Authorization with policies

+++ app/Http/Controllers/ThreadsController.php
@@ -6,6 +6,7 @@ use App\Channel;
 use App\Filters\ThreadFilters;
 use App\Thread;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ThreadsController extends Controller
 {
@@ -114,7 +115,7 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
-        //$thread->replies()->delete();
+        $this->authorize('delete', $thread);
 
         $thread->delete();
